moves global controls to own folder
